// src/modules/user-chats/types.ts
export enum ChatEnumTypes {
    GET_USER_CHATS_REQUEST = 'GET_USER_CHATS_REQUEST',
    GET_USER_CHATS_RESPONSE = 'GET_USER_CHATS_RESPONSE',
    GET_USER_CHATS_ERROR = 'GET_USER_CHATS_ERROR',

    GET_CHAT_REQUEST = 'GET_CHAT_REQUEST',
    GET_CHAT_RESPONSE = 'GET_CHAT_RESPONSE',
    GET_CHAT_ERROR = 'GET_CHAT_ERROR',

    OPEN_ITEM = 'OPEN_ITEM',
    OPEN_USER_CHAT = 'OPEN_USER_CHAT',

    GET_MESSAGES_REQUEST = 'GET_MESSAGES_REQUEST',
    GET_MESSAGES_RESPONSE = 'GET_MESSAGES_RESPONSE',
    GET_MESSAGES_ERROR = 'GET_MESSAGES_ERROR',

    POST_MESSAGE_REQUEST = 'POST_MESSAGE_REQUEST',
    POST_MESSAGE_RESPONSE = 'POST_MESSAGE_RESPONSE',
    POST_MESSAGE_ERROR = 'POST_MESSAGE_ERROR',

    SEND_MESSAGE = 'SEND_MESSAGE'
}

export interface PostMessage {
    message: string
    toUser: string
    chatId: string
}